Feat: Shrinks the created image considerably
Some notes.

This passes the hadolinter with one exception (DL3009),
as there is no reason for the image in the builder step to
shed extra weight as we end up with a small alpine based image
for the runner image.

The binary is statically compiled to make sure we do not get into
trouble wrt. alpines weirdo libc.